Fix unexpected surface pan when scroll view not at top
In some cases (e.g. drag the panel to `half`, then swipe slowly to
`full` and slightly overswipe so the content scroll view scrolls down a
bit) starting a drag in the grabber handle area would begin to shrink
the panel view, even though the inner scroll view was not in it's top
most state.

This may be the desired behavior for some use cases, but not mine.